move editor scene entry related stuff into separate mod
